% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electuk1920s.R
\docType{data}
\name{polres_scrape}
\alias{polres_scrape}
\alias{datalist}
\title{1920s election data}
\format{An object of class \code{list} of length 261.}
\usage{
datalist
}
\arguments{
\item{url}{Provide a URL link to one of political res's subpages}

\item{area}{Provide an area name for the function to retrieve data.}
}
\description{
This package will allow you to build a database, which is included as an RDS file in the repo, of 1920s election data.
This includes the 1922,1923,1924 and 1929 General Elections. This is done by webscraping HTML data from Wikipedia
and politicsresources.com. This contains data on the year, candidates, parties, constituencies, swing and margins of each seat.
I have also called on the Google Maps API to match the geographical data of the area. As such, it is advisable
for the user to obtain an API google maps key before starting.
}
\details{
To retrieve this data after building the dataset, the user can obtain the results using the filter
function in the tidyverse package by specifying the year and seat of concern.

There are two functions in this dataset package:
}
\keyword{data}
\keyword{election,}
\keyword{geographical}
\keyword{latitude,}
\keyword{longitude,}
\keyword{table}
